requirements.txt
aerospike==10.0.1
elasticsearch==8.6.2
Faker>=17
pandas>=1.5
psycopg2-binary==2.9.5
pymongo==4.3.3
PyYAML==6.0
redis==4.5.1

pytest >= 7